requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-groq
langchain-google-genai
langchain-ollama
sentence-transformers
pgvector
psycopg2-binary
groq
google-generativeai
openai
requests
beautifulsoup4
firecrawl-py
sounddevice
scipy
numpy
rank-bm25